Fixed the engine-on-strut fuel flow problem; Mouse/Keyboard zoom now working on the Hull Cameras; Added Warp to Maneuver Node; Translatron now accepts floating point numbers; Fixed several exceptions on loading; Obliterator fixes for 0.18.4;

+++ MuMechLib/MechJebModuleTranslatron.cs
@@ -99,7 +99,7 @@ namespace MuMech
 
                 if (GUILayout.Button("EXECUTE", sty, GUILayout.ExpandWidth(true)))
                 {
-                    core.trans_spd_act = Convert.ToInt16(trans_spd);
+                    core.trans_spd_act = Convert.ToSingle(trans_spd);
                     GUIUtility.keyboardControl = 0;
                 }
             }
